requirements.txt
absl-py==2.3.1
anyio==4.12.1
apache-beam==2.69.0
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
autograd==1.8.0
ax-platform==0.3.7
beartype==0.21.0
botorch==0.10.0
Bottleneck==1.5.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.0
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
debugpy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_debugpy_1754523443/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dnspython==2.7.0
docopt==0.6.2
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
fastavro==1.12.1
fasteners==0.20
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1755216263872/work
flatbuffers==25.12.19
fonttools==4.60.2
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1752608696525/work
gast==0.7.0
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1745509426225/work
google-pasta==0.2.0
gpytorch==1.11
grpcio==1.65.5
h11==0.16.0
h5py==3.14.0
hdfs==2.7.3
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface_hub==1.3.1
idna==3.11
ImageIO==2.37.2
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1754352890318/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1701831663892/work
ipywidgets==8.1.8
jaxtyping==0.2.36
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
joblib==1.5.3
jsonpickle==3.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1748333051527/work
jupyterlab_widgets==3.0.16
keras-nightly==3.10.0.dev2025052003
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
linear-operator==0.5.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1733219663979/work
matplotlib==3.9.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multipledispatch==1.0.0
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.15.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1698504735452/work
-e git+https://github.com/tonioeltopoquegira/neural-structural-optimization.git@c505d341e6784ba56c357db0d1273cf84272db86#egg=neural_structural_optimization
nlopt==2.9.1
numpy==1.26.4
objsize==0.7.1
opt_einsum==3.4.0
optree @ file:///Users/runner/miniforge3/conda-bld/optree_1753455083014/work
orjson==3.11.5
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1752081780253/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
plotly==6.5.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
proto-plus==1.27.0
protobuf==6.33.2
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1740663174295/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==21.0.0
pyarrow-hotfix==0.7
pybind11 @ file:///D:/bld/bld/rattler-build_pybind11_1755953378/work
pybind11-global @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pybind11-global_1755953338/work
pycparser==2.23
pydot==1.4.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyJWT==2.10.1
pymongo==4.15.5
pyparsing==3.3.1
pyre-extensions==0.0.32
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML==6.0.3
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1755799570317/work
redis==5.3.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
safetensors==0.7.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sortedcontainers==2.4.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
tb-nightly==2.20.0a20250717
tdqm==0.0.1
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf_nightly==2.21.0.dev20251029
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.24
torch==2.8.0
torchvision==0.23.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1754732067879/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typeguard==2.13.3
typer-slim==0.21.1
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3==2.6.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==2.0.1
xarray==2024.7.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.25.0
